+++ java-service-processor/src/main/java/internal/nbbrd/service/definition/ServiceDefinitionGenerator.java
@@ -27,6 +27,7 @@ import java.util.function.Consumer;
 import java.util.function.Predicate;
 import java.util.stream.Collector;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
 /**
@@ -64,14 +65,14 @@ class ServiceDefinitionGenerator {
     @lombok.NonNull
     List<LoadSorter> sorters;
 
-    public TypeSpec generate(boolean nested) {
+    public TypeSpec generateLoader(boolean nested) {
         String className = definition.resolveLoaderName().simpleName();
 
         TypeName quantifierType = getQuantifierType();
 
         FieldSpec sourceField = newSourceField();
-        MethodSpec spliteratorMethod = newSpliteratorMethod(sourceField);
-        MethodSpec doLoadMethod = newDoLoadMethod(spliteratorMethod, quantifierType);
+        Optional<FieldSpec> batchField = newBatchField();
+        MethodSpec doLoadMethod = newDoLoadMethod(sourceField, batchField, quantifierType);
         FieldSpec resourceField = newResourceField(doLoadMethod, quantifierType);
         MethodSpec getMethod = newGetMethod(resourceField, quantifierType);
 
@@ -79,8 +80,11 @@ class ServiceDefinitionGenerator {
                 .classBuilder(className)
                 .addJavadoc(getMainJavadoc())
                 .addModifiers(Modifier.PUBLIC, Modifier.FINAL)
-                .addField(sourceField)
-                .addMethod(spliteratorMethod)
+                .addField(sourceField);
+
+        batchField.ifPresent(result::addField);
+
+        result
                 .addMethod(doLoadMethod)
                 .addField(resourceField)
                 .addMethod(getMethod);
@@ -108,6 +112,29 @@ class ServiceDefinitionGenerator {
         return result.build();
     }
 
+    public Optional<TypeSpec> generateBatch(boolean nested) {
+        if (!definition.isBatch()) {
+            return Optional.empty();
+        }
+
+        String className = definition.resolveBatchName().simpleName();
+
+        TypeSpec.Builder result = TypeSpec
+                .interfaceBuilder(className)
+                .addModifiers(Modifier.PUBLIC)
+                .addMethod(MethodSpec
+                        .methodBuilder("getProviders")
+                        .addModifiers(Modifier.PUBLIC, Modifier.ABSTRACT)
+                        .returns(typeOf(Stream.class, definition.getServiceType()))
+                        .build());
+
+        if (nested) {
+            result.addModifiers(Modifier.STATIC).build();
+        }
+
+        return Optional.of(result.build());
+    }
+
     private CodeBlock getMainJavadoc() {
         return CodeBlock
                 .builder()
@@ -123,6 +150,8 @@ class ServiceDefinitionGenerator {
                 .add("<li>Name: $L</li>\n", definition.getLoaderName().isEmpty() ? "null" : definition.getLoaderName())
                 .add("<li>Backend: $L</li>\n", definition.getBackend().map(HasTypeMirror::getTypeName).orElse("null"))
                 .add("<li>Cleaner: $L</li>\n", definition.getCleaner().map(HasTypeMirror::getTypeName).orElse("null"))
+                .add("<li>Batch: $L</li>\n", definition.isBatch())
+                .add("<li>Batch name: $L</li>\n", definition.getBatchName().isEmpty() ? "null" : definition.getBatchName())
                 .add("</ul>\n")
                 .build();
     }
@@ -146,44 +175,7 @@ class ServiceDefinitionGenerator {
         return "null";
     }
 
-    private MethodSpec newSpliteratorMethod(FieldSpec sourceField) {
-        FieldSpec delegateField = FieldSpec
-                .builder(Iterator.class, "delegate")
-                .addModifiers(Modifier.FINAL)
-                .initializer("$N.iterator()", sourceField)
-                .build();
-
-        TypeSpec delegateClass = TypeSpec
-                .anonymousClassBuilder("$T.MAX_VALUE, 0", Long.class)
-                .superclass(typeOf(Spliterators.AbstractSpliterator.class, definition.getServiceType()))
-                .addField(delegateField)
-                .addMethod(MethodSpec
-                        .methodBuilder("tryAdvance")
-                        .addAnnotation(Override.class)
-                        .addModifiers(Modifier.PUBLIC)
-                        .addParameter(typeOf(Consumer.class, WildcardTypeName.supertypeOf(definition.getServiceType())), "action")
-                        .returns(TypeName.BOOLEAN)
-                        .addCode(CodeBlock
-                                .builder()
-                                .beginControlFlow("if ($N.hasNext())", delegateField)
-                                .addStatement("action.accept(($T) $N.next())", definition.getServiceType(), delegateField)
-                                .addStatement("return true")
-                                .endControlFlow()
-                                .build())
-                        .addStatement("return false")
-                        .build())
-                .build();
-
-        return MethodSpec
-                .methodBuilder("spliterator")
-                .addModifiers(Modifier.PRIVATE)
-                .addModifiers(getSingletonModifiers())
-                .returns(typeOf(Spliterator.class, definition.getServiceType()))
-                .addCode("return $L;\n", delegateClass)
-                .build();
-    }
-
-    private MethodSpec newDoLoadMethod(MethodSpec spliterator, TypeName quantifierType) {
+    private MethodSpec newDoLoadMethod(FieldSpec sourceField, Optional<FieldSpec> batchField, TypeName quantifierType) {
         return MethodSpec
                 .methodBuilder("doLoad")
                 .addModifiers(Modifier.PRIVATE)
@@ -193,14 +185,23 @@ class ServiceDefinitionGenerator {
                 .addStatement(CodeBlock
                         .builder()
                         .add("return ")
-                        .add(getPreprocessingCode(spliterator))
+                        .add(getPreprocessingCode(sourceField, batchField))
                         .add(getQuantifierCode())
                         .build())
                 .build();
     }
 
-    private CodeBlock getPreprocessingCode(MethodSpec spliterator) {
-        CodeBlock streamBlock = CodeBlock.of("$T.stream($N(), false)", StreamSupport.class, spliterator);
+    private CodeBlock newStreamBlock(FieldSpec iterable) {
+        return CodeBlock.of("$T.stream($N.spliterator(), false)", StreamSupport.class, iterable);
+    }
+
+    private CodeBlock getPreprocessingCode(FieldSpec sourceField, Optional<FieldSpec> batchField) {
+        CodeBlock sourceStream = newStreamBlock(sourceField);
+        Optional<CodeBlock> batchStream = batchField.map(this::newStreamBlock);
+
+        CodeBlock streamBlock = batchStream.isPresent()
+                ? CodeBlock.of("$T.concat($L, $L.flatMap(o -> o.getProviders()))", Stream.class, sourceStream, batchStream.get())
+                : sourceStream;
 
         return definition.getPreprocessor().isPresent()
                 ? getAdvancedPreprocessingCode(streamBlock, definition.getPreprocessor().get())
@@ -359,17 +360,28 @@ class ServiceDefinitionGenerator {
 
     private FieldSpec newSourceField() {
         return FieldSpec
-                .builder(Iterable.class, fieldName("source"))
+                .builder(typeOf(Iterable.class, definition.getServiceType()), fieldName("source"))
                 .addModifiers(Modifier.PRIVATE, Modifier.FINAL)
                 .addModifiers(getSingletonModifiers())
-                .initializer("$L", getBackendInitCode())
+                .initializer("$L", getBackendInitCode(definition.getServiceType()))
                 .build();
     }
 
-    private CodeBlock getBackendInitCode() {
+    private Optional<FieldSpec> newBatchField() {
+        return definition.isBatch()
+                ? Optional.of(FieldSpec
+                .builder(typeOf(Iterable.class, definition.resolveBatchName()), fieldName("batch"))
+                .addModifiers(Modifier.PRIVATE, Modifier.FINAL)
+                .addModifiers(getSingletonModifiers())
+                .initializer("$L", getBackendInitCode(definition.resolveBatchName()))
+                .build())
+                : Optional.empty();
+    }
+
+    private CodeBlock getBackendInitCode(ClassName serviceType) {
         return definition.getBackend().isPresent()
-                ? CodeBlock.of("$L.apply($T.class)", getInstantiatorCode(definition.getBackend().get()), definition.getServiceType())
-                : CodeBlock.of("$T.load($T.class)", ServiceLoader.class, definition.getServiceType());
+                ? CodeBlock.of("$L.apply($T.class)", getInstantiatorCode(definition.getBackend().get()), serviceType)
+                : CodeBlock.of("$T.load($T.class)", ServiceLoader.class, serviceType);
     }
 
     private FieldSpec newCleanerField() {