+++ java-service-processor/src/main/java/internal/nbbrd/service/definition/ServiceDefinitionProcessor.java
@@ -1,17 +1,17 @@
 /*
  * Copyright 2019 National Bank of Belgium
- * 
- * Licensed under the EUPL, Version 1.1 or - as soon they will be approved 
+ *
+ * Licensed under the EUPL, Version 1.1 or - as soon they will be approved
  * by the European Commission - subsequent versions of the EUPL (the "Licence");
  * You may not use this work except in compliance with the Licence.
  * You may obtain a copy of the Licence at:
- * 
+ *
  * http://ec.europa.eu/idabc/eupl
- * 
- * Unless required by applicable law or agreed to in writing, software 
+ *
+ * Unless required by applicable law or agreed to in writing, software
  * distributed under the Licence is distributed on an "AS IS" basis,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Licence for the specific language governing permissions and 
+ * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
 package internal.nbbrd.service.definition;
@@ -21,10 +21,7 @@ import com.squareup.javapoet.JavaFile;
 import com.squareup.javapoet.TypeSpec;
 import internal.nbbrd.service.ProcessorUtil;
 import internal.nbbrd.service.Unreachable;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
+
 import javax.annotation.processing.AbstractProcessor;
 import javax.annotation.processing.Processor;
 import javax.annotation.processing.RoundEnvironment;
@@ -32,16 +29,20 @@ import javax.annotation.processing.SupportedAnnotationTypes;
 import javax.lang.model.SourceVersion;
 import javax.lang.model.element.Modifier;
 import javax.lang.model.element.TypeElement;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
- *
  * @author Philippe Charles
  */
 @org.openide.util.lookup.ServiceProvider(service = Processor.class)
 @SupportedAnnotationTypes({
-    "nbbrd.service.ServiceDefinition",
-    "nbbrd.service.ServiceFilter",
-    "nbbrd.service.ServiceSorter"
+        "nbbrd.service.ServiceDefinition",
+        "nbbrd.service.ServiceFilter",
+        "nbbrd.service.ServiceSorter"
 })
 public final class ServiceDefinitionProcessor extends AbstractProcessor {
 
@@ -67,7 +68,7 @@ public final class ServiceDefinitionProcessor extends AbstractProcessor {
         Map<ClassName, List<LoadDefinition>> definitionsByTopLevel = data.getDefinitions()
                 .stream()
                 .filter(checker::checkDefinition)
-                .collect(Collectors.groupingBy(definition -> definition.resolveLoaderName().topLevelClassName()));
+                .collect(Collectors.groupingBy(definition -> definition.getServiceType().topLevelClassName()));
 
         Map<ClassName, List<LoadFilter>> filtersByService = data.getFilters()
                 .stream()
@@ -85,20 +86,46 @@ public final class ServiceDefinitionProcessor extends AbstractProcessor {
     }
 
     private void generate(ClassName topLevel, List<ServiceDefinitionGenerator> generators) {
-        TypeSpec typeSpec = isNotNested(topLevel, generators)
-                ? generators.get(0).generate(false)
-                : generateNested(topLevel, generators);
-        ProcessorUtil.write(processingEnv, JavaFile.builder(topLevel.packageName(), typeSpec).build());
+        if (isNotNested(topLevel, generators)) {
+            generateNotNested(generators.get(0));
+        } else {
+            generateNested(topLevel, generators);
+        }
     }
 
     private boolean isNotNested(ClassName topLevel, List<ServiceDefinitionGenerator> generators) {
-        return generators.size() == 1 && topLevel.equals(generators.get(0).getDefinition().resolveLoaderName());
+        return generators.size() == 1 && topLevel.equals(generators.get(0).getDefinition().getServiceType());
     }
 
-    private TypeSpec generateNested(ClassName topLevel, List<ServiceDefinitionGenerator> generators) {
-        return TypeSpec.classBuilder(topLevel)
+    private void generateNotNested(ServiceDefinitionGenerator generator) {
+        String loaderPackage = generator.getDefinition().resolveLoaderName().packageName();
+        TypeSpec loaderClass = generator.generateLoader(false);
+        writeFile(loaderPackage, loaderClass);
+
+        String batchPackage = generator.getDefinition().resolveBatchName().packageName();
+        Optional<TypeSpec> batchClass = generator.generateBatch(false);
+        batchClass.ifPresent(batchSpec -> writeFile(batchPackage, batchSpec));
+    }
+
+    private void generateNested(ClassName topLevel, List<ServiceDefinitionGenerator> generators) {
+        List<TypeSpec> nestedLoaders = generators.stream().map(o -> o.generateLoader(true)).collect(Collectors.toList());
+        TypeSpec loaderClass = TypeSpec.classBuilder(ClassName.bestGuess(topLevel.canonicalName() + "Loader"))
                 .addModifiers(Modifier.PUBLIC, Modifier.FINAL)
-                .addTypes(generators.stream().map(o -> o.generate(true)).collect(Collectors.toList()))
+                .addTypes(nestedLoaders)
                 .build();
+        writeFile(topLevel.packageName(), loaderClass);
+
+        List<TypeSpec> nestedBatches = generators.stream().map(o -> o.generateBatch(true)).filter(Optional::isPresent).map(Optional::get).collect(Collectors.toList());
+        if (!nestedBatches.isEmpty()) {
+            TypeSpec batchClass = TypeSpec.classBuilder(ClassName.bestGuess(topLevel.canonicalName() + "Batch"))
+                    .addModifiers(Modifier.PUBLIC, Modifier.FINAL)
+                    .addTypes(nestedBatches)
+                    .build();
+            writeFile(topLevel.packageName(), batchClass);
+        }
+    }
+
+    private void writeFile(String loaderPackage, TypeSpec loaderClass) {
+        ProcessorUtil.write(processingEnv, JavaFile.builder(loaderPackage, loaderClass).build());
     }
 }